% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slipper.R
\name{slipper}
\alias{slipper}
\title{Bootstrap a function that returns a single number}
\usage{
slipper(df, expr, B = 100)
}
\arguments{
\item{df}{A data frame}

\item{expr}{A bare function that operates on one or more variables from the data frame.}

\item{B}{the number of bootstrap samples to draw}
}
\value{
out A data frame with the values, whether they come from the observed data or the bootstrapped data.
}
\description{
Takes a data frame, a function that returns a single number,
and a number of replicates and returns a data frame with the
function calculated on the observed sample and on B bootstrap
samples.
}
\examples{

# Boostrap the mean of the mpg variable in the mtcars data set
slipper(mtcars,mean(mpg),B=100)

# Bootstrap the mean of the mpg variable with piping
mtcars \%>\% slipper(mean(mpg),B=100)

# Calculate a confidence interval using the quantiles method
mtcars \%>\% slipper(mean(mpg),B=100) \%>\%
 filter(type="bootstrap") \%>\%
 summarize(ci_low = quantile(value,0.025),
           ci_high = quantile(value,0.975))


}

