% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slipper_ci.R
\name{slipper_ci}
\alias{slipper_ci}
\title{Bootstrap a function that returns a single number and
return a non-parametric confidence interval.}
\usage{
slipper_ci(df, expr, B = 100, lower = 0.025, upper = 0.975)
}
\arguments{
\item{df}{A data frame}

\item{expr}{A an expression with a function that operates on one or more variables from the data frame.}

\item{B}{the number of bootstrap samples to draw}

\item{lower}{the lower percentile for the confidence interval (default 2.5%)}

\item{upper}{the upper percentile for the confidence interval (default 97.5%)}
}
\value{
confint A vector with the lower bound and upper bound
}
\description{
Takes a data frame, a function that returns a single number,
and a number of replicates and returns a data frame with the
function calculated on the observed sample and on B bootstrap
samples.
}
\examples{

# Boostrap the mean of the mpg variable in the mtcars data set and get a confidence
# interval 
slipper_ci(mtcars,mean(mpg),B=100)

# Bootstrap the mean of the mpg variable with piping
mtcars \%>\% slipper_ci(mean(mpg),B=100)



}
